styling changes to data

+++ packages/frontend/app/components/guild/modal/data/Reset.tsx
@@ -6,34 +6,36 @@ export default function Reset({ data, setGuild, setData }: DataComponentProps) {
   const reset = async (event: any) => {
     event.preventDefault();
 
-    const res = await fetch("/api/discord/webhook", {
-      method: "DELETE",
-      body: JSON.stringify({
-        id: data.body.webhookId,
-      }),
-    });
+    const confirm = event.target.confirm.value;
 
-    const updatedGuild = await fetch("/api/dynamo/reset", {
-      method: "POST",
-      body: JSON.stringify({
-        guild: { guild_id: data.body.guildId },
-      }),
-    }).then((result) => result.json());
+    if (confirm === "delete") {
+      const res = await fetch("/api/discord/webhook", {
+        method: "DELETE",
+        body: JSON.stringify({
+          id: data.body.webhookId,
+        }),
+      });
 
-    setGuild(updatedGuild);
-    setData({ command: "", body: {} });
+      const updatedGuild = await fetch("/api/dynamo/reset", {
+        method: "POST",
+        body: JSON.stringify({
+          guild: { guild_id: data.body.guildId },
+        }),
+      }).then((result) => result.json());
+
+      setGuild(updatedGuild);
+      setData({ command: "", body: {} });
+    }
   };
 
   return (
     <div data-testid="Reset">
       <Form onSubmit={reset}>
         <Form.Group className="mb-3">
-          <Row>
-            <Col>
-              WARNING: this will completely remove all history of this server,
-              this action is not reversable
-            </Col>
-          </Row>
+          <Form.Control
+            name="confirm"
+            placeholder="Please type 'delete' to reset"
+          />
           <br></br>
           <Button
             type="submit"
